Split core/base.py into two files
