+++ superduperdb/core/dataset.py
@@ -4,7 +4,8 @@ import typing as t
 
 import numpy
 
-from superduperdb.core.base import Component, Artifact
+from superduperdb.core.artifact import Artifact
+from superduperdb.core.component import Component
 from superduperdb.core.documents import Document
 from superduperdb.datalayer.mongodb.query import Find
 import dataclasses as dc
